% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_application.R
\name{create_appDir}
\alias{create_appDir}
\title{Create an appDir for shiny::runApp and rsconnect::deployApp}
\usage{
create_appDir(datafile = NULL)
}
\arguments{
\item{datafile}{The data set file used by the app}
}
\value{
The created directory
}
\description{
Create an appDir for shiny::runApp and rsconnect::deployApp
}
