% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamiskTabellverk-package.R
\docType{package}
\name{dynamiskTabellverk}
\alias{dynamiskTabellverk}
\alias{dynamiskTabellverk-package}
\title{dynamiskTabellverk}
\description{
Package including functions used by dynamisk tabellverk Helse-Nord RHF
}
