% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filtrerUt}
\alias{filtrerUt}
\title{Filtrere ut data fra datasett}
\usage{
filtrerUt(tabell, fane, verdi, aar, bo, beh, behandlingsniva, alder, kjonn,
  hastegrad2, hdg, icd10, fag)
}
\arguments{
\item{tabell}{Datasettet som skal filtreres.}

\item{fane, rad, kol, verdi, aar, bo, beh, behandlingsniva, alder, kjonn, hastegrad2, hdg, icd10}{Filteringsvalg.}
}
\description{
Filtrerer ut data fra datasett som ikke skal tabuleres
}
