% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/just_overf.R
\name{just_overf}
\alias{just_overf}
\title{Title}
\usage{
just_overf(input, output, session, colnames)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{colnames}{}
}
\value{

}
\description{
Title
}
