% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttonModules.R
\name{just_overf_server}
\alias{just_overf_server}
\title{Module for the "adjust for transfers" button}
\usage{
just_overf_server(input, output, session, colnames)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{colnames}{Column names in the data}

\item{id}{Shiny id}
}
\value{

}
\description{
Module for the "adjust for transfers" button
}
