% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttonModules.R
\name{just_overf_ui}
\alias{just_overf_ui}
\title{Module for the "adjust for transfers" button}
\usage{
just_overf_ui(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{

}
\description{
Module for the "adjust for transfers" button
}
