% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttonModules.R
\name{justOverf}
\alias{justOverf}
\title{Server side of the "adjust for transfers" button module}
\usage{
justOverf(input, output, session, colnames)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{colnames}{Column names in the data}
}
\value{

}
\description{
Server side of the "adjust for transfers" button module
}
