% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttonModules.R
\name{justOverfUI}
\alias{justOverfUI}
\title{Module for the "adjust for transfers" button}
\usage{
justOverfUI(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{

}
\description{
Module for the "adjust for transfers" button
}
