% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{lagHjelpetekst}
\alias{lagHjelpetekst}
\title{Title}
\usage{
lagHjelpetekst(tab, rad, kol, verdi, aar, bo, beh, prosent, behandlingsniva,
  alder, kjonn, hastegrad2, forenkling)
}
\arguments{
\item{tab, rad, kol, verdi, aar, bo, beh, prosent, behandlingsniva, alder, kjonn, hastegrad2, forenkling}{Beskrivelse}
}
\value{
Tekst
}
\description{
Title
}
