% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sortdataset}
\alias{sortdataset}
\title{Data frame used to define sorting of values in table}
\format{
A data frame with 3 variables
\describe{
  \item{variable}{Variable to sort}
  \item{value}{Value of variable}
  \item{sort}{Integer for sorting}
}
}
\usage{
sortdataset
}
\description{
Data frame used to define sorting of values in table
}
\keyword{datasets}
