% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_application.R
\name{submit_application}
\alias{submit_application}
\title{Submit the application to shinyapp.io}
\usage{
submit_application(datasett = NULL, name = "experimental")
}
\arguments{
\item{datasett}{The data set file (.RData) to be loaded into the application.
The absolute path has to be given}

\item{name}{The appName of the deployed shiny application}
}
\description{
Submit the application to shinyapp.io
}
