% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata2}
\alias{testdata2}
\title{Some test data 2}
\format{
An object of class \code{data.frame} with 8969 rows and 23 columns.
}
\usage{
testdata2
}
\description{
Some test data 2
}
\keyword{datasets}
