% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxModules.R
\name{checkbox}
\alias{checkbox}
\alias{aar_server}
\alias{behandlingsniva_server}
\alias{hastegrad1_server}
\alias{hastegrad2_server}
\alias{kjonn_server}
\alias{alder_server}
\alias{prosent_server}
\alias{snitt_server}
\alias{keep_names_server}
\title{Module for the year checkbox}
\usage{
aar_server(id, pickable)

behandlingsniva_server(id, pickable, colnames)

hastegrad1_server(id, pickable, colnames)

hastegrad2_server(id, pickable, colnames)

kjonn_server(id, pickable, colnames)

alder_server(id, pickable, colnames)

prosent_server(id, pickable)

snitt_server(id)

keep_names_server(id)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\description{
Module for the year checkbox

Module for the behandlingsniva checkbox

Module for the hastegrad1 checkbox

Module for the hastegrad2 checkbox

Module for the kjonn checkbox

Module for the alder checkbox

Module for the knappProsent checkbox

Module for the knappSnitt checkbox

Module for the keep_name checkbox
}
