% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{definer_valg_kol}
\alias{definer_valg_kol}
\title{Define the possible selections}
\usage{
definer_valg_kol(col_names, valgnr)
}
\arguments{
\item{valgnr}{Hvilket valg}

\item{datasett}{Datasett}
}
\value{
valg_en valg_to valg_tre
}
\description{
Based on the content of the data
}
