% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{definerValgKol}
\alias{definerValgKol}
\title{Title}
\usage{
definerValgKol(datasett, valgnr)
}
\arguments{
\item{datasett}{}

\item{valgnr}{}
}
\value{
valg_en valg_to valg_tre
}
\description{
Title
}
\examples{
TBA
}
