% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttonModules.R
\name{just_overf_server}
\alias{just_overf_server}
\title{Server side of the "adjust for transfers" button module}
\usage{
just_overf_server(input, output, session, colnames)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{colnames}{Column names in the data}
}
\value{

}
\description{
Server side of the "adjust for transfers" button module
}
