% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{makeDataTabell}
\alias{makeDataTabell}
\title{Title}
\usage{
makeDataTabell(inpDatasett, fane, rad, kol, verdi, aar, bo, beh,
  behandlingsniva, alder, kjonn, hastegrad2, prosent, forenkling, keepNames,
  snitt, hdg, icd10, fag)
}
\arguments{
\item{inpDatasett}{a}

\item{fane}{a}

\item{rad}{a}

\item{kol}{a}

\item{verdi}{a}

\item{aar}{a}

\item{bo}{a}

\item{beh}{a}

\item{behandlingsniva}{a}

\item{alder}{a}

\item{kjonn}{a}

\item{hastegrad2}{a}

\item{prosent}{a}

\item{forenkling}{a}

\item{keepNames}{a}

\item{snitt}{a}

\item{hdg}{a}

\item{icd10}{a}
}
\value{
pivot
}
\description{
Title
}
