% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{makeDataTabell}
\alias{makeDataTabell}
\title{Title}
\usage{
makeDataTabell(inpDatasett, fane, rad, kol, verdi, aar, bo, beh,
  behandlingsniva, alder, kjonn, hastegrad2, prosent, forenkling, keepNames,
  snitt, hdg, icd10)
}
\arguments{
\item{inpDatasett}{}

\item{fane}{}

\item{rad}{}

\item{kol}{}

\item{verdi}{}

\item{aar}{}

\item{bo}{}

\item{beh}{}

\item{behandlingsniva}{}

\item{alder}{}

\item{kjonn}{}

\item{hastegrad2}{}

\item{prosent}{}

\item{forenkling}{}

\item{keepNames}{}

\item{snitt}{}

\item{hdg}{}

\item{icd10}{}
}
\value{
pivot
}
\description{
Title
}
\examples{
TBA
}
