% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{makeDataTabell}
\alias{makeDataTabell}
\title{Make a pivot table}
\usage{
makeDataTabell(inpDatasett, fane, verdier, keepNames, snitt)
}
\arguments{
\item{inpDatasett}{The dataset that is going to be tabulated}

\item{fane}{The active tab (for filtering; effect if "dogn", "dag" or "poli")}

\item{verdier}{Consist of data specific filter values, such as "rader", "kolonner", "aar", "bo", "beh" etc.}

\item{keepNames}{Tabulate all names in first row}

\item{snitt}{Add average/sums to the table}
}
\value{
pivot Pivot table
}
\description{
Make a pivot table
}
