% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_application.R
\name{submit_application}
\alias{submit_application}
\title{Submit the application to shinyapp.io}
\usage{
submit_application(datasett = NULL)
}
\arguments{
\item{datasett}{The data set to be loaded into the application}
}
\description{
Submit the application to shinyapp.io
}
